Working on ChunkInfection - Currently somewhat broken.
Working on new CursorQueue system for ChunkInfection.

The cursor queue will be used so that we can limit how many top down cursors, as well as surface cursors created from topdown ones, are spawned in at any given time. Still not 100% set on how everything will work yet but the next push should be when everything is complete.

For now this is just an experiment. Lots more todo, but adding to github as a backup for now.